+++ kotonebot/client/implements/windows/__init__.py
@@ -1,1 +1,2 @@
 from .windows import *
+from .windows_native import *